add category name to category file
The category file consists of <rule,id> tuples,
where "id" is the numeric id of the desired category.
IDs aren't intuitive to use, so this patch extends the
existing functionality with an id-to-name lookup
allowing <rule,name> to be used in the tuples.
This is similar to the protocols file,
where rules end in a protocol name.